% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPutPrice}
\alias{getPutPrice}
\title{Compute Up-and-Out Put Option price with a barrier}
\usage{
getPutPrice(
  Expiry = 0.75,
  Strike = 150,
  Spot = 140,
  b = 160,
  Vol = 0.24,
  r = 0.07,
  NumberOfPaths = 10000L,
  LengthOfPath = 10000L
)
}
\arguments{
\item{Expiry}{Time to Maturity}

\item{Strike}{Strike price}

\item{Spot}{Spot price}

\item{b}{Barrier value}

\item{Vol}{Volatility}

\item{r}{Risk-free rate}

\item{NumberOfPaths}{Number of Monte-Carlo paths}

\item{LengthOfPath}{Number of Monte-Carlo paths}
}
\description{
Compute Up-and-Out Put Option price with a barrier
}
\examples{
getPutPrice(0.75, 150, 140, 160, 0.24, 0.07)
}
